chore: moving file to correct folder to aligns with convention
